error handle

+++ src/modules/customer/customer.service.ts
@@ -1,5 +1,7 @@
+import { StatusCodes } from 'http-status-codes';
 import { PrismaClient } from '@prisma/client';
 import { ICustomer } from './customer.interface';
+import AppError from '../../errors/AppError';
 
 const prisma = new PrismaClient();
 const createCustomerIntoDB = async (customerInfo: ICustomer) => {
@@ -15,6 +17,13 @@ const getAllCustomerFromDB = async () => {
 };
 
 const getSingleCustomerFromDB = async (id: string) => {
+  const isExist = await prisma.customer.findUnique({
+    where: { customerId: id },
+  });
+
+  if (!isExist) {
+    throw new AppError(StatusCodes.NOT_FOUND, 'Customer not found');
+  }
   const customers = prisma.customer.findUnique({
     where: {
       customerId: id,
@@ -27,6 +36,13 @@ const updateCustomerInfoInDB = async (
   id: string,
   payload: Partial<ICustomer>
 ) => {
+  const isExist = await prisma.customer.findUnique({
+    where: { customerId: id },
+  });
+
+  if (!isExist) {
+    throw new AppError(StatusCodes.NOT_FOUND, 'Customer not found');
+  }
   const updatedCustomer = await prisma.customer.update({
     where: {
       customerId: id,
@@ -38,6 +54,14 @@ const updateCustomerInfoInDB = async (
 };
 
 const deleteCustomerFromDB = async (id: string) => {
+  const isExist = await prisma.customer.findUnique({
+    where: { customerId: id },
+  });
+
+  if (!isExist) {
+    throw new AppError(StatusCodes.NOT_FOUND, 'Customer not found');
+  }
+
   await prisma.customer.delete({
     where: {
       customerId: id,
